// src/models/_fndds-models/db.d.ts
export interface dataSrcAttribute {
  dataSrcId: string;
  authors?: string;
  title?: string;
  year?: string;
  journal?: string;
  volCity?: string;
  issueState?: string;
  startPage?: string;
  endPage?: string;
}

export interface dataSrcInstance extends Sequelize.Instance<dataSrcAttribute>, dataSrcAttribute {
}